requirements.txt
Package                              Version
------------------------------------ ---------------
matplotlib                           3.4.3
matplotlib-inline                    0.1.3
mmcv-full                            1.1.2
PyYAML                               6.0
pyzmq                                19.0.2
ruamel.yaml                          0.17.4
ruamel.yaml.clib                     0.2.6
scikit-image                         0.19.1
scipy                                1.7.2
torch                                1.5.0
torchvision                          0.6.0a0+82fd1c8
